Symbexec: add get_block
